Add database migration for Sequel schema.
This database should be painfully simple. We are really only storing
references so that we are able to build a proper URL to the GitHub
repository.

There is obvious room for changes here such as if a repository moves,
gets deleted, etc. That logic may need to be something that is sketched
out later.