Update models, API with Users and Gears.
This includes a new namespace which allows queries to happen for the
Users. There have been changes to the models that now draw associations
and thus allow for easy eager loading.

In addition there is now webhook endpoint for GitHub integration. The
thought here is that new repositories after being added will have,
through Octokit, a hook added that will notify when releases happen.